# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ennote ennoworker API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/graphs/{graphID}:
    parameters:
      - { name: graphID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getGraph
      responses:
        "200":
          description: Parsed Graph draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateGraph
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GraphMutation" }
      responses:
        "200":
          description: Semantically updated Graph draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /v1/sessions/{sessionID}/graph-runs:
    parameters:
      - { name: sessionID, in: path, required: true, schema: { type: string } }
    get:
      operationId: listSessionGraphRuns
      summary: List the Session's Graph Runs, newest first
      responses:
        "200":
          description: Session Graph Runs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { type: array, items: { $ref: "#/components/schemas/RunAgentFlow" } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionID}/graph-runs/{runID}:
    parameters:
      - { name: sessionID, in: path, required: true, schema: { type: string } }
      - { name: runID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getSessionGraphRun
      summary: One Graph Run with task checkpoints and the frozen flow version
      responses:
        "200":
          description: Graph Run detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [run, nodes]
                        properties:
                          run: { $ref: "#/components/schemas/RunAgentFlow" }
                          nodes:
                            type: array
                            items: { $ref: "#/components/schemas/RunAgentFlowNode" }
                          flowVersion: { type: integer }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: cancelSessionGraphRun
      summary: Cancel a running Graph Run
      responses:
        "200":
          description: The (possibly now terminal) Graph Run
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/RunAgentFlow" } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sessions/{sessionID}/graph-runs/{runID}/resume:
    parameters:
      - { name: sessionID, in: path, required: true, schema: { type: string } }
      - { name: runID, in: path, required: true, schema: { type: string } }
    post:
      operationId: resumeSessionGraphRun
      summary: Resume a cancelled or failed Graph Run
      responses:
        "200":
          description: The reset Graph Run (pending)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/RunAgentFlow" } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/graphs/{graphID}/builder:
    parameters:
      - { name: graphID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getGraphBuilderThread
      responses:
        "200":
          description: Persistent Builder thread and pending proposal
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphBuilderThread" } }
  /v1/graphs/{graphID}/builder/messages:
    parameters:
      - { name: graphID, in: path, required: true, schema: { type: string } }
    post:
      operationId: sendGraphBuilderMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [modelProfileId, instruction]
              properties:
                modelProfileId: { type: string }
                instruction: { type: string, minLength: 1, maxLength: 65536 }
      responses:
        "201":
          description: Builder response and validated proposal
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphBuilderThread" } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/graphs/{graphID}/builder/proposals/{proposalID}/apply:
    parameters:
      - { name: graphID, in: path, required: true, schema: { type: string } }
      - { name: proposalID, in: path, required: true, schema: { type: string } }
    post:
      operationId: applyGraphBuilderProposal
      responses:
        "200":
          description: Proposal applied atomically to graph.yaml
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
...
    RunAgentFlow:
      type: object
      properties:
        runId: { type: string }
        sessionId: { type: string }
        projectId: { type: string }
        flowVersionId: { type: string }
        manifestDigest: { type: string }
        state: { type: string, enum: [pending, running, completed, failed, cancelled, convergence_exceeded, budget_exceeded] }
        totalTokensUsed: { type: integer }
        terminalReason: { type: string }
        inputs: { type: object, additionalProperties: true }
        createdAt: { type: string }
        updatedAt: { type: string }
        finishedAt: { type: string }
    RunAgentFlowNode:
      type: object
      properties:
        runId: { type: string }
        taskIndex: { type: integer }
        handle: { type: string }
        roleVersionId: { type: string }
        skillDigests: { type: array, items: { type: string } }
        goalDigest: { type: string }
        goalText: { type: string }
        budget: { type: object, additionalProperties: true }
        terminalState: { type: string, enum: [pending, running, completed, failed, blocked, cancelled, interrupted] }
        outputRef: { type: object, additionalProperties: true }
        childRunId: { type: string }
        errorCode: { type: string }
        createdAt: { type: string }
        finishedAt: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
...
    ThinkingEffort:
      type: string
      enum: [default, low, medium, high]
...
    GraphTask:
      type: object
      additionalProperties: false
      required: [name, goal]
      description: Role-backed Tasks declare role only; inline Tasks declare model with optional thinking and skills. The modes are mutually exclusive and enforced by the Worker.
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string, pattern: "^(local|global)/[a-z][a-z0-9_-]{1,63}$" }
        model: { type: string, pattern: "^[^/]+/[^/]+$" }
        thinking: { $ref: "#/components/schemas/ThinkingEffort" }
        skills: { type: array, uniqueItems: true, items: { type: string, pattern: "^(local|global)/[a-z][a-z0-9_-]{1,63}$" } }
        goal: { type: string, minLength: 1 }
        writes: { type: array, uniqueItems: true, items: { type: string } }
        budget:
          type: object
          additionalProperties: false
          properties: { tokens: { type: integer, minimum: 1 } }
    GraphDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, id, name, tasks, graph]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        id: { type: string }
        name: { type: string }
        description: { type: string }
        tasks: { type: object, additionalProperties: { $ref: "#/components/schemas/GraphTask" } }
        graph:
          type: object
          additionalProperties: { type: array, uniqueItems: true, items: { type: string } }
    GraphSummary:
      type: object
      additionalProperties: false
      required: [id, name, path, latestVersion]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        digest: { type: string }
        error: { type: string }
        latestVersion: { type: integer, minimum: 0 }
    GraphDetail:
      allOf:
        - $ref: "#/components/schemas/GraphSummary"
        - type: object
          required: [digest, document]
          properties:
            digest: { type: string }
            document: { $ref: "#/components/schemas/GraphDocument" }
    GraphMutation:
      type: object
      additionalProperties: false
      required: [expectedDigest]
      properties:
        expectedDigest: { type: string }
        name: { type: string }
        description: { type: string }
        task:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id: { type: string }
            value: { $ref: "#/components/schemas/GraphTask" }
        dependencies:
          type: object
          additionalProperties: false
          required: [taskId, depends]
          properties:
            taskId: { type: string }
            depends: { type: array, uniqueItems: true, items: { type: string } }
    GraphBuilderMessage:
      type: object
      additionalProperties: false
      required: [id, graphId, ordinal, role, content, createdAt]
      properties:
        id: { type: string }
        graphId: { type: string }
        ordinal: { type: integer, minimum: 1 }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    GraphBuilderOperation:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, enum: [upsert_task, delete_task, set_dependencies, update_graph] }
        taskId: { type: string }
        task: { $ref: "#/components/schemas/GraphTask" }
        depends: { type: array, items: { type: string } }
        name: { type: string }
        description: { type: string }
    GraphBuilderProposal:
      type: object
      additionalProperties: false
      required: [id, graphId, baseDigest, operations, summary, status, diagnostics, createdAt]
      properties:
        id: { type: string }
        graphId: { type: string }
        baseDigest: { type: string }
        operations: { type: array, items: { $ref: "#/components/schemas/GraphBuilderOperation" } }
        summary: { type: string }
        status: { type: string, enum: [pending, applied, superseded] }
        diagnostics: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    GraphBuilderThread:
      type: object
      additionalProperties: false
      required: [graphId, messages]
      properties:
        graphId: { type: string }
        modelProfileId: { type: string }
        messages: { type: array, items: { $ref: "#/components/schemas/GraphBuilderMessage" } }
        proposal: { $ref: "#/components/schemas/GraphBuilderProposal" }
...